requirements.txt
fastapi==0.111.0
RPi.GPIO
rpi-lgpio